join criteria in one controller
