+++ src/main/java/com/fundacionjala/convertor/controller/SearchController.java
@@ -1,6 +1,7 @@
 package com.fundacionjala.convertor.controller;
 
 import com.fundacionjala.convertor.model.FileSearcher;
+import com.fundacionjala.convertor.view.AdvancedSearchVideo;
 import com.fundacionjala.convertor.view.SearchView;
 
 import java.awt.event.ActionEvent;
@@ -18,14 +19,16 @@ public class SearchController implements ActionListener {
     private SearchView searchView;
     private FileSearcher fileSearcher;
     private SearchCriteria searchCriteria;
+    private AdvancedSearchVideo advancedSearchVideo;
 
 
     /**
      * @param searchView   searchView
      * @param fileSearcher fileSearcher
      */
-    public SearchController(final SearchView searchView, final FileSearcher fileSearcher) {
+    public SearchController(final SearchView searchView, final AdvancedSearchVideo advancedSearchVideo,final FileSearcher fileSearcher) {
         this.searchView = searchView;
+        this.advancedSearchVideo = advancedSearchVideo;
         this.fileSearcher = fileSearcher;
         //this.searchView.btnSearch.addActionListener(this);
         searchCriteria = new SearchCriteria();
@@ -40,6 +43,9 @@ public class SearchController implements ActionListener {
         searchCriteria.setPath(searchView.getTxtPath().getText());
         searchCriteria.setExt(searchView.getCmbFileFormat().getSelectedItem().toString());
         searchCriteria.setSize(searchView.getCmbSize().getSelectedItem().toString());
+        searchCriteria.setFps( Integer.parseInt(advancedSearchVideo.getCmbFps().getSelectedItem().toString()));
+        searchCriteria.setAspectRatio(advancedSearchVideo.getCmbAspectRatio().getSelectedItem().toString());
+        searchCriteria.setResolution(advancedSearchVideo.getCmbResolution().getSelectedItem().toString());
 
     }
