+++ src/main/java/com/fundacionjala/convertor/view/MainView.java
@@ -101,8 +101,9 @@ public class MainView extends JFrame implements ActionListener {
         if (e.getSource() == btnSearch) {
 
             SearchView searchView = new SearchView(this);
+            AdvancedSearchVideo advancedSearchVideo = new AdvancedSearchVideo();
             FileSearcher fileSearcher = new FileSearcher();
-            SearchController searchController = new SearchController(searchView, fileSearcher);
+            SearchController searchController = new SearchController(searchView, advancedSearchVideo, fileSearcher);
             // searchView.pack();
             searchView.setVisible(true);
             if (searchView.getIsOk()) {
